Card Payment templates
npm install react-square-web-payments-sdk

npm install square

SquareForm.js: Front end. Creates payment form when making credit or debit purchases. For now this is a template and will have more features added very soon. Also generates a unique token for each purchase for security.

pay.js: Back end. This file will receive the incoming payment request and then call the checkout API method. Currently, this is still a template and will be filled in with the necessary features. This will also check whether a payment is successfully completed or not and will correctly respond by POST the order information to the DB.